+++ ComputerServiceShopSolution/CSOS.Core/DTO/OfferDto/OfferUpdateRequest.cs
@@ -1,10 +1,11 @@
 ﻿using CSOS.Core.DTO.DtoContracts;
 using Microsoft.AspNetCore.Http;
 
-namespace CSOS.Core.DTO.Requests
+namespace CSOS.Core.DTO.OfferDto
 {
-    public class OfferAddRequest : IOfferDeliveryDto, IOfferImageDto
+    public class OfferUpdateRequest : IOfferDeliveryDto, IOfferImageDto
     {
+        public int Id { get; set; }
         public string ProductName { get; set; } = null!;
         public string Description { get; set; } = null!;
         public int SelectedProductCategory { get; set; }
@@ -15,6 +16,7 @@ namespace CSOS.Core.DTO.Requests
         public decimal Price { get; set; }
         public int StockQuantity { get; set; }
         public List<int> SelectedOtherDeliveries { get; set; } = null!;
+        public List<string>? ImagesToDelete { get; set; } = new List<string>();
         public int? SelectedParcelLocker { get; set; }
     }
 }
