+++ ComputerServiceShopSolution/CSOS.Core/DTO/OfferDto/UserOffersResponseDto.cs
@@ -1,4 +1,4 @@
-﻿namespace CSOS.Core.DTO.Responses.Offers
+﻿namespace CSOS.Core.DTO.OfferDto
 {
     public class UserOffersResponseDto
     {
@@ -7,9 +7,10 @@
         public string ProductName { get; set; } = null!;
         public string ProductCategory { get; set; } = null!;
         public string ProductCondition { get; set; } = null!;
-        public bool ProductStatus { get; set; }
+        public bool IsOfferPrivate { get; set; }
 
         public int StockQuantity;
+        
         public decimal Price;
         public DateTime DateCreated { get; set; }
     }
